\name{optimStrat-package}
\alias{optimStrat-package}
\alias{optimStrat}
\docType{package}
\title{optimStrat}
\description{OptimStrat is a package intended to assist in the choice of the sample strategy to implement in a survey. It compares five strategies having into account the information available in an auxiliary variable and two superpopulation models, called working and true models.}
\details{
The package includes a web-based application where the user can compare five sampling strategies in order to determine which one to implement in a survey.

The package also includes a function to perform the comparison mentioned above, as well as functions for stratifying the auxiliary variable and calculations of the variance of Stratified Simple Random Sampling and Pareto \eqn{\pi}ps.
}
\author{
Edgar Bueno
}
\references{
Bueno, E. (2018). \emph{A Comparison of Stratified Simple Random Sampling and Probability Proporional-to-size Sampling.} Research Report, Department of Statistics, Stockholm University 2018:6. \url{http://gauss.stat.su.se/rr/RR2018_6.pdf}.
}
\keyword{ package }
\keyword{ survey }
